+++ ScriptingDefinitionEditor/IDEScriptingDefinitionEditor.h
@@ -50,7 +50,7 @@
 - (void)changeTextSize:(id)arg1;
 - (void)scriptingDefinition_makeTextSmaller:(id)arg1;
 - (void)scriptingDefinition_makeTextLarger:(id)arg1;
-@property int viewMode;
+@property long long viewMode;
 - (void)selectDocumentLocations:(id)arg1;
 - (void)observeValueForKeyPath:(id)arg1 ofObject:(id)arg2 change:(id)arg3 context:(void *)arg4;
 - (void)_updateSelectedDefinitions;
